chore(typescript-estree): rename internal module name
