+++ packages/typescript-estree/src/semantic-or-syntactic-errors.ts
